add tests and documentation for sales_management
